# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Workforce scheduling API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ── Users ─────────────────────────────────────────────────────────────────
...
  /companies/{id}:
    get:
      operationId: getCompany
      tags: [companies]
      summary: Get company details
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: updateCompany
      tags: [companies]
      summary: Update company (admin only)
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCompanyRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"

  # ── Workplaces ────────────────────────────────────────────────────────────
...
  /availability:
    get:
      operationId: listAvailability
      tags: [availability]
      summary: List availability records
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Availability"
    post:
      operationId: createAvailability
      tags: [availability]
      summary: Submit availability
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AvailabilityRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Availability"
  /availability/{id}:
    get:
      operationId: getAvailability
      tags: [availability]
      summary: Get availability record
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Availability"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: updateAvailability
      tags: [availability]
      summary: Update availability record
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AvailabilityRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Availability"

  # ── Leave Requests ────────────────────────────────────────────────────────
...
  /invitations:
    get:
      operationId: listInvitations
      tags: [invitations]
      summary: List invitations (admin/manager)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invitation"
    post:
      operationId: createInvitation
      tags: [invitations]
      summary: Create an invitation (admin/manager)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvitationRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invitation"
  /invitations/{id}:
    get:
      operationId: getInvitation
      tags: [invitations]
      summary: Get an invitation
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invitation"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteInvitation
      tags: [invitations]
      summary: Cancel an invitation
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # ── Auth ────────────────────────────────────────────────────────────────
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        name:
          type: string
        companyName:
          type: string
          description: Provide to create a new company (you become admin)
        invitationToken:
          type: string
          description: Provide to join an existing company via invitation
      required: [email, password, name]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/AuthUser"
      required: [token, user]
    AuthUser:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
        companyId:
          type: integer
          nullable: true
      required: [id, email, name, role]
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
        status:
          type: string
        phone:
          type: string
          nullable: true
        hourlyRate:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        companyId:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, email, name, role, status]
...
    Company:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        status:
          type: string
        plan:
          type: string
        logoUrl:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        timezone:
          type: string
        weekStartDay:
          type: integer
        overtimeThreshold:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, name, status, plan, timezone, weekStartDay, overtimeThreshold]
    UpdateCompanyRequest:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        timezone:
          type: string
        weekStartDay:
          type: integer
          minimum: 0
          maximum: 6
        overtimeThreshold:
          type: string
        logoUrl:
          type: string
          nullable: true

    # ── Workplace ────────────────────────────────────────────────────────────
...
    Availability:
      type: object
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        companyId:
          type: integer
        weekStart:
          type: string
          description: ISO date string e.g. 2026-07-06
        slots:
          type: object
          description: "Flexible JSON e.g. { monday: ['09:00-17:00'] }"
          additionalProperties: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, employeeId, companyId, weekStart, slots]
    AvailabilityRequest:
      type: object
      properties:
        weekStart:
          type: string
        slots:
          type: object
          additionalProperties: true
        notes:
          type: string
          nullable: true
        employeeId:
          type: integer
          description: Admin/manager only — submit on behalf of an employee
      required: [weekStart, slots]

    # ── Leave Request ─────────────────────────────────────────────────────────
...
    Invitation:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
          nullable: true
        email:
          type: string
        role:
          type: string
          enum: [admin, manager, employee]
        status:
          type: string
          enum: [pending, accepted, expired]
        invitedBy:
          type: integer
          nullable: true
        token:
          type: string
        expiresAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, email, role, status, token]
    InvitationRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        role:
          type: string
          enum: [admin, manager, employee]
        expiresAt:
          type: string
          format: date-time
      required: [email]